Updated the code for EmailController.java
This code now displays and updates the selected file to send through the email for the GetSelectFile() action (lines 62-63), replaces the password variable with user input for Google Security App Password reasons (line 99). Also new Error message for password field is added (line 129), and Error message for trying to send the email with the wrong password is modified (lines 116-117).