Updated code for EmailForm.java
Modified the form which adds a new label and text field for the app password, added a new label for the file name, and modified the "Email Address" label to "Recipient Email Address".  Added getters and setters for the password and file name related items. Added a new public void EmailFormWithFileName() function which refreshes the form with the path name of the selected item from the dialog box menu. The file name refreshes to No File Selected after an email is sent or the user returns to the main menu.